requirements.txt
-i https://pypi.org/simple
bleach==6.1.0; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
contourpy==1.2.1; python_version >= '3.9'
cycler==0.12.1; python_version >= '3.8'
filelock==3.15.4; python_version >= '3.8'
fonttools==4.53.1; python_version >= '3.8'
fsspec==2024.6.1; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
jinja2==3.1.4; python_version >= '3.7'
kaggle==1.6.14
kiwisolver==1.4.5; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
matplotlib==3.9.1; python_version >= '3.9'
mpmath==1.3.0
networkx==3.3; python_version >= '3.10'
numpy==2.0.0; python_version >= '3.9'
nvidia-cublas-cu12==12.1.3.1; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-cupti-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-nvrtc-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-runtime-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cudnn-cu12==8.9.2.26; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cufft-cu12==11.0.2.54; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-curand-cu12==10.3.2.106; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusolver-cu12==11.4.5.107; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusparse-cu12==12.1.0.106; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nccl-cu12==2.20.5; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nvjitlink-cu12==12.5.82; python_version >= '3'
nvidia-nvtx-cu12==12.1.105; platform_system == 'Linux' and platform_machine == 'x86_64'
packaging==24.1; python_version >= '3.8'
pillow==10.4.0; python_version >= '3.8'
pyparsing==3.1.2; python_full_version >= '3.6.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-slugify==8.0.4; python_version >= '3.7'
requests==2.32.3; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sympy==1.13.0; python_version >= '3.8'
text-unidecode==1.3
torch==2.3.1; python_full_version >= '3.8.0'
torchvision==0.18.1; python_version >= '3.8'
tqdm==4.66.4; python_version >= '3.7'
triton==2.3.1; python_version < '3.12' and platform_system == 'Linux' and platform_machine == 'x86_64'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.2; python_version >= '3.8'
webencodings==0.5.1
